# FrontEnd/Ra3d Chat System/demo ui/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ChatServerMVC",
    "version": "1.0"
  },
  "paths": {
...
    "/api/rooms/room": {
      "post": {
        "tags": [
          "Room"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/rooms/directMessage": {
      "post": {
        "tags": [
          "Room"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDMRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDMRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDMRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateDMRequest": {
        "type": "object",
        "properties": {
          "secondUser": {
            "type": "string",
            "nullable": true
          },
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuidByteArrayValueTuple"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateRoomRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "creator": {
            "type": "string",
            "format": "uuid"
          },
          "users": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "encryptionKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuidByteArrayValueTuple"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GuidByteArrayValueTuple": {
        "type": "object",
        "additionalProperties": false
      },
...
    }
  },
...
}